requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
black==24.4.2
certifi==2024.7.4
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
etils==1.5.2
exceptiongroup==1.2.2
executing==2.0.1
filelock==3.15.4
flax==0.8.5
frozenlist==1.4.1
fsspec==2024.6.1
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipython==8.18.1
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-pinecone==0.1.2
langchain-text-splitters==0.2.2
langsmith==0.1.93
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
opt-einsum==3.3.0
optax==0.2.3
orbax-checkpoint==0.5.22
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pinecone==4.0.0
pinecone-client==4.1.2
pinecone-plugin-interface==0.0.7
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.27.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
stack-data==0.6.3
sympy==1.13.1
tenacity==8.5.0
tensorstore==0.1.63
threadpoolctl==3.5.0
tokenizers==0.19.1
tomli==2.0.1
toolz==0.12.1
torch==2.3.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
yarl==1.9.4
zipp==3.19.2
